break up components to make it easier to move things around
