+++ src/locales/en/main.json
@@ -3653,6 +3653,7 @@
     "refresh": "Refresh templates",
     "edit": "Edit",
     "view": "View",
+    "templateDetails": "Template Details",
     "done": "Done",
     "publish": "Publish",
     "publishing": "Publishing...",
