+++ src/platform/marketplace/components/PublishTemplateWizard.vue
@@ -1,305 +1,72 @@
 <template>
-  <div class="flex flex-col gap-4">
-    <div v-if="!readOnly" class="flex items-center gap-2">
-      <div class="ml-auto flex gap-2 text-sm text-muted">
-        <span
-          v-for="(step, idx) in steps"
-          :key="step.id"
-          :class="
-            cn(
-              'rounded-full px-3 py-1 select-none',
-              currentStep === idx + 1
-                ? 'text-highlight-foreground bg-highlight'
-                : 'bg-base-background'
-            )
-          "
-        >
-          {{ step.label }}
-        </span>
-      </div>
-    </div>
+  <BaseModalLayout size="md">
+    <template #header>
+      <h2>
+        {{
+          readOnly
+            ? $t('marketplace.templateDetails')
+            : $t('marketplace.publishToMarketplace')
+        }}
+      </h2>
+    </template>
 
-    <!-- Step 1: Details + Preview split view -->
-    <div
-      v-if="currentStep === 1"
-      data-testid="step-details"
-      class="grid grid-cols-1 gap-6 lg:grid-cols-2"
-    >
-      <div class="flex flex-col gap-4">
-        <div class="flex flex-col gap-1">
-          <label for="publish-title" class="text-sm font-medium">
-            {{ $t('marketplace.title') }}
-          </label>
-          <input
-            v-if="readOnly"
-            id="publish-title"
-            :value="form.title"
-            data-testid="input-title"
-            type="text"
-            readonly
-            class="border-border rounded-md border bg-muted/30 px-3 py-2 text-sm"
-          />
-          <input
-            v-else
-            id="publish-title"
-            v-model="form.title"
-            data-testid="input-title"
-            type="text"
-            class="border-border rounded-md border bg-base-background px-3 py-2 text-sm"
-          />
-        </div>
-
-        <div class="flex flex-col gap-1">
-          <label for="publish-description" class="text-sm font-medium">
-            {{ $t('marketplace.description') }}
-          </label>
-          <textarea
-            v-if="readOnly"
-            id="publish-description"
-            :value="form.description"
-            data-testid="input-description"
-            rows="4"
-            readonly
-            class="border-border rounded-md border bg-muted/30 px-3 py-2 text-sm"
-          />
-          <textarea
-            v-else
-            id="publish-description"
-            v-model="form.description"
-            data-testid="input-description"
-            rows="4"
-            class="border-border rounded-md border bg-base-background px-3 py-2 text-sm"
-          />
-        </div>
-
-        <div class="flex flex-col gap-1">
-          <label for="publish-short-description" class="text-sm font-medium">
-            {{ $t('marketplace.shortDescription') }}
-          </label>
-          <input
-            v-if="readOnly"
-            id="publish-short-description"
-            :value="form.shortDescription"
-            data-testid="input-short-description"
-            type="text"
-            readonly
-            class="border-border rounded-md border bg-muted/30 px-3 py-2 text-sm"
-          />
-          <input
-            v-else
-            id="publish-short-description"
-            v-model="form.shortDescription"
-            data-testid="input-short-description"
-            type="text"
-            class="border-border rounded-md border bg-base-background px-3 py-2 text-sm"
-          />
-        </div>
-
-        <div class="flex flex-col gap-1">
-          <label for="publish-license" class="text-sm font-medium">
-            {{ $t('marketplace.license') }}
-          </label>
-          <SingleSelect
-            v-if="!readOnly"
-            id="publish-license"
-            v-model="form.license"
-            :label="$t('marketplace.license')"
-            :options="licenseOptions"
-            data-testid="input-license"
-            size="md"
-          />
-          <input
-            v-else
-            id="publish-license"
-            :value="licenseLabel"
-            data-testid="input-license"
-            type="text"
-            readonly
-            class="border-border rounded-md border bg-muted/30 px-3 py-2 text-sm"
-          />
-        </div>
-
-        <div class="flex flex-col gap-1">
-          <label class="text-sm font-medium">
-            {{ $t('marketplace.tags') }}
-          </label>
-          <TagInputWithAutocomplete
-            v-if="!readOnly"
-            v-model="form.tags"
-            :placeholder="$t('marketplace.tagsPlaceholder')"
-          />
-          <div
-            v-else
-            class="flex flex-wrap gap-1"
-            data-testid="input-tags-readonly"
-          >
-            <span
-              v-for="tag in form.tags"
-              :key="tag"
-              class="rounded-full bg-muted/50 px-2 py-0.5 text-xs"
-            >
-              {{ tag }}
-            </span>
-          </div>
-        </div>
-      </div>
-
-      <div class="flex flex-col gap-2">
-        <p class="text-xs text-muted">
-          {{ $t('marketplace.previewDescription') }}
-        </p>
-        <div
-          class="flex w-full justify-center rounded-lg border border-border-default bg-base-background py-5"
-        >
-          <div class="w-full max-w-72">
-            <MarketplaceTemplatePreviewCard
-              :title="form.title"
-              :short-description="form.shortDescription"
-              :description="form.description"
-              :license-label="licenseLabel"
-              :tags="form.tags"
+    <template #content>
+      <div class="flex min-h-0 flex-1 flex-col">
+        <div class="min-h-0 flex-1 overflow-y-auto">
+          <div class="flex flex-col gap-4">
+            <PublishTemplateWizardStepDetails
+              v-if="currentStep === 1"
+              v-model:title="form.title"
+              v-model:description="form.description"
+              v-model:short-description="form.shortDescription"
+              v-model:license="form.license"
+              v-model:tags="form.tags"
+              :read-only="readOnly"
               :thumbnail-url="thumbnailUrl"
-            >
-              <template v-if="!readOnly" #thumbnail-placeholder>
-                <div
-                  class="relative flex size-full flex-col items-center justify-center gap-2 rounded-lg border-2 border-dashed transition-colors"
-                  :class="
-                    cn(
-                      'cursor-pointer border-border-default bg-dialog-surface hover:border-muted-foreground',
-                      isOverThumbnailDrop && 'border-muted-foreground'
-                    )
-                  "
-                  data-testid="preview-thumbnail-placeholder"
-                  data-handles-file-drop
-                  @click="thumbnailFileInputRef?.click()"
-                  @dragover.prevent="handleThumbnailDragOver"
-                  @dragleave="handleThumbnailDragLeave"
-                  @drop.prevent="handleThumbnailDrop"
-                >
-                  <input
-                    ref="thumbnailFileInputRef"
-                    type="file"
-                    accept="image/*"
-                    class="hidden"
-                    data-testid="input-thumbnail-file"
-                    @change="handleThumbnailFileSelect"
-                  />
-                  <div
-                    v-if="isUploadingThumbnail"
-                    class="absolute inset-0 z-10 flex items-center justify-center rounded-lg bg-base-background/80"
-                  >
-                    <i
-                      class="icon-[lucide--loader-circle] size-10 animate-spin text-muted"
-                      aria-hidden
-                    />
-                  </div>
-                  <i class="icon-[lucide--image] size-10 text-muted" />
-                  <span class="text-xs text-muted">
-                    {{ $t('marketplace.dropThumbnailHere') }}
-                  </span>
-                </div>
-              </template>
-            </MarketplaceTemplatePreviewCard>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Step 2: Submit -->
-    <div
-      v-if="currentStep === 2"
-      data-testid="step-submit"
-      class="flex flex-col gap-4"
-    >
-      <div
-        v-if="submitted"
-        class="border-success bg-success/10 rounded-lg border p-4 text-sm"
-      >
-        {{ $t('marketplace.submitted') }}
-      </div>
-      <div v-else class="flex flex-col gap-2 text-sm">
-        <p>{{ $t('marketplace.previewDescription') }}</p>
-        <div class="flex justify-center">
-          <div class="w-full max-w-72">
-            <MarketplaceTemplatePreviewCard
-              :title="form.title"
-              :short-description="form.shortDescription"
-              :description="form.description"
+              :license-options="licenseOptions"
               :license-label="licenseLabel"
-              :tags="form.tags"
+              :is-uploading-thumbnail="isUploadingThumbnail"
+              @thumbnail-selected="(file) => void processThumbnailFile(file)"
+            />
+
+            <PublishTemplateWizardStepSubmit
+              v-if="currentStep === 2"
+              :form="form"
+              :submitted="submitted"
+              :license-label="licenseLabel"
               :thumbnail-url="thumbnailUrl"
             />
           </div>
         </div>
+
+        <PublishTemplateWizardFooter
+          class="shrink-0"
+          :error="error"
+          :read-only="readOnly"
+          :steps="steps"
+          :current-step="currentStep"
+          :submitted="submitted"
+          :is-pending-review="isPendingReview"
+          :can-advance="canAdvance"
+          :is-publishing="isPublishing"
+          @close="handleClose"
+          @back="handleBack"
+          @next="handleNext"
+          @submit="handleSubmit"
+        />
       </div>
-    </div>
-
-    <!-- Error display -->
-    <div
-      v-if="error"
-      class="bg-danger/10 text-danger rounded-md px-3 py-2 text-sm"
-    >
-      {{ error }}
-    </div>
-
-    <!-- Navigation buttons -->
-    <div v-if="readOnly" class="flex justify-end pt-2">
-      <Button data-testid="btn-done" @click="onClose?.()">
-        {{ $t('marketplace.done') }}
-      </Button>
-    </div>
-    <div v-else class="flex justify-between pt-2">
-      <Button
-        v-if="currentStep > 1"
-        data-testid="btn-back"
-        variant="secondary"
-        @click="handleBack"
-      >
-        {{ $t('marketplace.back') }}
-      </Button>
-      <div v-else />
-
-      <div class="flex gap-2">
-        <Button
-          v-if="currentStep < 2"
-          data-testid="btn-next"
-          :disabled="!canAdvance"
-          @click="handleNext"
-        >
-          {{ $t('marketplace.next') }}
-        </Button>
-        <Button
-          v-if="currentStep === 2 && !submitted && isPendingReview"
-          data-testid="btn-done"
-          @click="handleDone"
-        >
-          {{ $t('marketplace.done') }}
-        </Button>
-        <Button
-          v-if="currentStep === 2 && !submitted && !isPendingReview"
-          data-testid="btn-submit"
-          :disabled="isPublishing"
-          @click="handleSubmit"
-        >
-          {{
-            isPublishing
-              ? $t('marketplace.submitting')
-              : $t('marketplace.submitForReview')
-          }}
-        </Button>
-      </div>
-    </div>
-  </div>
+    </template>
+  </BaseModalLayout>
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, reactive, ref, watch } from 'vue'
+import { computed, onMounted, provide, reactive, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-import TagInputWithAutocomplete from '@/components/input/TagInputWithAutocomplete.vue'
-import SingleSelect from '@/components/input/SingleSelect.vue'
-import Button from '@/components/ui/button/Button.vue'
-import MarketplaceTemplatePreviewCard from '@/platform/marketplace/components/MarketplaceTemplatePreviewCard.vue'
+import BaseModalLayout from '@/components/widget/layout/BaseModalLayout.vue'
+import PublishTemplateWizardFooter from '@/platform/marketplace/components/PublishTemplateWizardFooter.vue'
+import PublishTemplateWizardStepDetails from '@/platform/marketplace/components/PublishTemplateWizardStepDetails.vue'
+import PublishTemplateWizardStepSubmit from '@/platform/marketplace/components/PublishTemplateWizardStepSubmit.vue'
 import { createGraphThumbnail } from '@/renderer/core/thumbnail/graphThumbnailRenderer'
 import type {
   LicenseType,
@@ -308,8 +75,7 @@ import type {
 import { LICENSE_TYPES } from '@/platform/marketplace/apiTypes'
 import { useMarketplacePublishing } from '@/platform/marketplace/composables/useMarketplacePublishing'
 import { useToastStore } from '@/platform/updates/common/toastStore'
-import { extractFilesFromDragEvent, hasImageType } from '@/utils/eventUtils'
-import { cn } from '@/utils/tailwindUtil'
+import { OnCloseKey } from '@/types/widgetTypes'
 
 const {
   onClose,
@@ -320,6 +86,8 @@ const {
   initialTemplate?: MarketplaceTemplate
   readOnly?: boolean
 }>()
+
+provide(OnCloseKey, onClose ?? (() => {}))
 
 const { t } = useI18n()
 
@@ -367,8 +135,6 @@ const submitted = ref(false)
 const thumbnailUrl = ref<string | null>(null)
 const pendingThumbnailFile = ref<File | null>(null)
 const isUploadingThumbnail = ref(false)
-const isOverThumbnailDrop = ref(false)
-const thumbnailFileInputRef = ref<HTMLInputElement | null>(null)
 
 const isPendingReview = computed(
   () => initialTemplate?.status === 'pending_review'
@@ -512,29 +278,8 @@ async function handleSubmit() {
   }
 }
 
-function handleDone() {
+function handleClose() {
   onClose?.()
-}
-
-function hasImageDropData(e: DragEvent): boolean {
-  if (!e.dataTransfer) return false
-  if (e.dataTransfer.files?.length) {
-    return Array.from(e.dataTransfer.files).some(hasImageType)
-  }
-  return ['text/uri-list', 'text/x-moz-url'].some((t) =>
-    e.dataTransfer!.types.includes(t)
-  )
-}
-
-function handleThumbnailDragOver(e: DragEvent) {
-  if (!hasImageDropData(e)) return
-  e.preventDefault()
-  e.dataTransfer!.dropEffect = 'copy'
-  isOverThumbnailDrop.value = true
-}
-
-function handleThumbnailDragLeave() {
-  isOverThumbnailDrop.value = false
 }
 
 function fileToDataUrl(file: File): Promise<string> {
@@ -567,30 +312,5 @@ async function processThumbnailFile(imageFile: File) {
   } else {
     pendingThumbnailFile.value = imageFile
   }
-}
-
-async function handleThumbnailDrop(e: DragEvent) {
-  isOverThumbnailDrop.value = false
-  e.preventDefault()
-  e.stopPropagation()
-  const files = await extractFilesFromDragEvent(e)
-  const imageFile = files.find(hasImageType)
-  if (!imageFile) {
-    useToastStore().addAlert(t('marketplace.thumbnailUploadError'))
-    return
-  }
-  await processThumbnailFile(imageFile)
-}
-
-function handleThumbnailFileSelect(e: Event) {
-  const input = e.target as HTMLInputElement
-  const file = input.files?.[0]
-  if (!file) return
-  if (!hasImageType(file)) {
-    useToastStore().addAlert(t('marketplace.thumbnailUploadError'))
-    return
-  }
-  void processThumbnailFile(file)
-  input.value = ''
 }
 </script>
