+++ src/platform/marketplace/composables/usePublishDialog.ts
@@ -1,11 +1,13 @@
 import type { MarketplaceTemplate } from '@/platform/marketplace/apiTypes'
 
 import PublishTemplateWizard from '@/platform/marketplace/components/PublishTemplateWizard.vue'
+import { useDialogService } from '@/services/dialogService'
 import { useDialogStore } from '@/stores/dialogStore'
 
 const DIALOG_KEY = 'publish-to-marketplace'
 
 export function usePublishDialog() {
+  const dialogService = useDialogService()
   const dialogStore = useDialogStore()
 
   function show(options?: {
@@ -17,22 +19,13 @@ export function usePublishDialog() {
       options?.onClose?.()
       hide()
     }
-    dialogStore.showDialog({
+    dialogService.showLayoutDialog({
       key: DIALOG_KEY,
-      title: options?.readOnly ? 'Template Details' : 'Publish to Marketplace',
       component: PublishTemplateWizard,
       props: {
         onClose: handleClose,
         initialTemplate: options?.initialTemplate,
         readOnly: options?.readOnly
-      },
-      dialogComponentProps: {
-        class: 'min-w-[600px]',
-        onClose: handleClose,
-        pt: {
-          header: 'p-5!',
-          content: 'p-5! overflow-y-hidden!'
-        }
       }
     })
   }
